[CI] Add native build image for linux-riscv64
This is used as part of the riseproject-dev out-of-tree CI [1][2], to get us closer to having native builds of PyTorch on RISC-V.

[1] https://github.com/risepytorch-dev/pytorch
[2] https://github.com/pytorch/pytorch/blob/01bcbb4c83936f38c54a35abcef28a052bc25f83/.github/allowlist.yml#L40

+++ .ci/docker/common/install_cache.sh
@@ -3,29 +3,36 @@
 set -ex
 
 install_ubuntu() {
-  echo "Installing pkg-config and libssl-dev"
-  apt-get update && apt-get install -y pkg-config libssl-dev curl
-  echo "Installing rust"
-  curl https://sh.rustup.rs -sSf | sh -s -- -y
-  echo "Checking out sccache repo"
-  git clone https://github.com/mozilla/sccache -b v0.13.0
-  cd sccache
-  echo "Patch dist build on aarch64"
-  sed -i '/all(target_os = "linux", target_arch = "x86_64"),/{ p; s/x86_64/aarch64/; }' src/bin/sccache-dist/main.rs
-  echo "Building sccache"
-  . "$HOME/.cargo/env" && cargo build --release --features="dist-client dist-server"
-  cp target/release/sccache /opt/cache/bin
-  cp target/release/sccache-dist /opt/cache/bin
-  echo "Cleaning up"
-  cd ..
-  rm -rf sccache
-  rustup self uninstall -y
-  apt-get remove -y pkg-config libssl-dev
-  apt-get autoclean && apt-get clean
+  if [[ "$BUILD_ENVIRONMENT" == *riscv64* ]]; then
+    # Download from official mozilla/sccache releases
+    curl -fsSL https://github.com/mozilla/sccache/releases/download/v0.15.0/sccache-v0.15.0-riscv64gc-unknown-linux-musl.tar.gz \
+      | tar -xvzf - -C /opt/cache/bin --strip-components 1 --wildcards 'sccache-*-riscv64gc-unknown-linux-musl/sccache'
+  else
+    [[ "$BUILD_ENVIRONMENT" == *aarch64* ]] || (echo "we are not building for aarch64"; exit 1)
+    echo "Installing pkg-config and libssl-dev"
+    apt-get update && apt-get install -y pkg-config libssl-dev curl
+    echo "Installing rust"
+    curl https://sh.rustup.rs -sSf | sh -s -- -y
+    echo "Checking out sccache repo"
+    git clone https://github.com/mozilla/sccache -b v0.13.0
+    cd sccache
+    echo "Patch dist build on aarch64"
+    sed -i '/all(target_os = "linux", target_arch = "x86_64"),/{ p; s/x86_64/aarch64/; }' src/bin/sccache-dist/main.rs
+    echo "Building sccache"
+    . "$HOME/.cargo/env" && cargo build --release --features="dist-client dist-server"
+    cp target/release/sccache /opt/cache/bin
+    cp target/release/sccache-dist /opt/cache/bin
+    echo "Cleaning up"
+    cd ..
+    rm -rf sccache
+    rustup self uninstall -y
+    apt-get remove -y pkg-config libssl-dev
+    apt-get autoclean && apt-get clean
 
-  echo "Downloading old sccache binary from S3 repo for PCH builds"
-  curl --retry 3 https://s3.amazonaws.com/ossci-linux/sccache -o /opt/cache/bin/sccache-0.2.14a
-  chmod 755 /opt/cache/bin/sccache-0.2.14a
+    echo "Downloading old sccache binary from S3 repo for PCH builds"
+    curl --retry 3 https://s3.amazonaws.com/ossci-linux/sccache -o /opt/cache/bin/sccache-0.2.14a
+    chmod 755 /opt/cache/bin/sccache-0.2.14a
+  fi
 }
 
 install_binary() {
